[FEAT] Validation on ProductController&CategoryController [FEAT] API Product

+++ app/Http/Controllers/ProductController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Product;
 use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
@@ -17,7 +17,7 @@ class ProductController extends Controller
     {
         $products = Product::all();
         $categories = Category::all();
-        return view('product.index', compact('products','categories'),['judul'=>'Product']);
+        return view('product.index', compact('products', 'categories'), ['judul' => 'Product']);
     }
 
     /**
@@ -26,7 +26,7 @@ class ProductController extends Controller
     public function create()
     {
         $categories = Category::all();
-        return view('product.create',compact('categories'),['judul'=>'Product']);
+        return view('product.create', compact('categories'), ['judul' => 'Product']);
     }
 
     /**
@@ -34,19 +34,32 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        $validate = Validator::make($request->all(), [
+            'name' => 'required|string|unique:products,name',
+            'category' => 'required',
+            'product_code' => 'required|unique:products',
+            'description' => 'required',
+            'price' => 'required',
+            'unit' => 'required',
+            'stock' => 'required',
+            'image' => 'image|mimes:jpeg,png,jpg|max:2048',
+        ]);
+        if ($validate->fails()) {
+            return redirect()->back()->withErrors($validate->errors())->withInput();
+        }
         $imageName = time() . '.' . $request->image->extension();
 
         Storage::putFileAs('image', $request->image, $imageName);
 
         $products = Product::create([
-            'name'=>$request->name,
-            'category_id'=>$request->category,
-            'product_code'=>$request->product_code,
-            'description'=>$request->description,
-            'price'=>$request->price,
-            'unit'=>$request->unit,
-            'stock'=>$request->stock,
-            'image'=>$imageName
+            'name' => $request->name,
+            'category_id' => $request->category,
+            'product_code' => $request->product_code,
+            'description' => $request->description,
+            'price' => $request->price,
+            'unit' => $request->unit,
+            'stock' => $request->stock,
+            'image' => $imageName,
         ]);
 
         return redirect('/product');
@@ -67,7 +80,7 @@ class ProductController extends Controller
     {
         $products = Product::find($id);
         $categories = Category::all();
-        return view('product.edit',compact('products','categories'),['judul'=>'Product']);
+        return view('product.edit', compact('products', 'categories'), ['judul' => 'Product']);
     }
 
     /**
@@ -75,37 +88,52 @@ class ProductController extends Controller
      */
     public function update($id, Request $request)
     {
+
+        $validate = Validator::make($request->all(), [
+            'name' => 'required|string',
+            'category' => 'required',
+            'product_code' => 'required|unique:products,product_code,'. $id.',id',
+            'description' => 'required',
+            'price' => 'required',
+            'unit' => 'required',
+            'stock' => 'required',
+            'image' => 'image|mimes:jpeg,png,jpg|max:2048',
+        ]);
+
+        if ($validate->fails()) {
+            return redirect()->back()->withErrors($validate->errors())->withInput();
+        }
+
         if ($request->hasFile('image')) {
-        $old_image = Product::find($id)->image;
+            $old_image = Product::find($id)->image;
 
-        Storage::delete('image/'.$old_image);
+            Storage::delete('image/' . $old_image);
 
-        $imageName = time() . '.' . $request->image->extension();
+            $imageName = time() . '.' . $request->image->extension();
 
-        Storage::putFileAs('image', $request->image, $imageName);
+            Storage::putFileAs('image', $request->image, $imageName);
 
-        $products = Product::where( 'id',$id)->update([
-        'name'=>$request->name,
-        'category_id'=>$request->category,
-        'product_code'=>$request->product_code,
-        'description'=>$request->description,
-        'price'=>$request->price,
-        'unit'=>$request->unit,
-        'stock'=>$request->stock,
-        'image'=>$imageName
-        ]);
-}
-    else {
-        $products = Product::where( 'id',$id)->update([
-            'name'=>$request->name,
-            'category_id'=>$request->category,
-            'product_code'=>$request->product_code,
-            'description'=>$request->description,
-            'price'=>$request->price,
-            'unit'=>$request->unit,
-            'stock'=>$request->stock,
-        ]);
-    }
+            $products = Product::where('id', $id)->update([
+                'name' => $request->name,
+                'category_id' => $request->category,
+                'product_code' => $request->product_code,
+                'description' => $request->description,
+                'price' => $request->price,
+                'unit' => $request->unit,
+                'stock' => $request->stock,
+                'image' => $imageName,
+            ]);
+        } else {
+            $products = Product::where('id', $id)->update([
+                'name' => $request->name,
+                'category_id' => $request->category,
+                'product_code' => $request->product_code,
+                'description' => $request->description,
+                'price' => $request->price,
+                'unit' => $request->unit,
+                'stock' => $request->stock,
+            ]);
+        }
         return redirect('/product');
     }
